Smol state improvements
Mostly in the process of re-familiarizing myself with this code, noticed
a couple things that could be improved around the actions and state.

Specifically, the actions to update various parts of the boards,
depended on the cell getting passed through the action, which, in
theory, could create a scenario where the cell/action state is somehow
out of sync with the board.  Logically the way it's built, it should
never happen, but it's a good mental exercise in re-thinking what the
source of truth SHOULD be.